% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraph2D.R
\name{createGraph2D}
\alias{createGraph2D}
\title{Creates an undirected graph from a biological sequence}
\usage{
createGraph2D(matrix, numSeqMRNA, nameMeasure)
}
\arguments{
\item{matrix}{matrix of the measure for the creation of two-dimensional graph}

\item{numSeqMRNA}{number of mRNA sequences}

\item{nameMeasure}{name of the measure to put in the title of the graph}
}
\value{
Returns the non-directed graph formed through the sequence
}
\description{
A function that from a biological sequence generates a graph
not directed having as vertices words, this being able to have its size
parameter set by the' word 'parameter. The connections between words depend
of the' step 'parameter that indicates the next connection to be formed.
}
\details{

}
\examples{

}
\seealso{

}
\author{
Eric Augusto Ito
}
