% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraph3D.R
\name{createGraph3D}
\alias{createGraph3D}
\title{Creates a three-dimensional chart between two measurements and the threshold}
\usage{
createGraph3D(matrix1, matrix2, numSeqMRNA, numSeqLNCRNA, nameMeasure1,
  nameMeasure2)
}
\arguments{
\item{matrix1}{matrix of the first measure for the creation of the three-dimensional chart}

\item{matrix2}{matrix of the second measure for the creation of the three-dimensional chart}

\item{numSeqMRNA}{number of mRNA sequences}

\item{numSeqLNCRNA}{number of lncRNA sequences}

\item{nameMeasure1}{name of the first measure to put in the title of the graph}

\item{nameMeasure2}{name of the second measure to put in the title of the graph}
}
\value{
Show a graphic three-dimensional
}
\description{
For an analysis of each measure, the createGraph2D () function was created in order to visualize the behavior of each measurement in relation to the threshold. This function creates a graph (Measure x Measure x Threshold) from two arrays, mRNA sequences are given the blue color, the lncRNA sequences are given a red color. In cases where there is a third class this will be given the green color
}
\author{
Eric Augusto Ito
}
