% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNet.R
\name{createNet}
\alias{createNet}
\title{Creates an untargeted graph from a biological sequence}
\usage{
createNet(word, step, sequence)
}
\arguments{
\item{word}{This integer parameter decides the size of the word that will be formed}

\item{step}{It is the integer parameter that decides the step that will be taken to make a new connection}

\item{sequence}{It is a vector that represents the sequence}
}
\value{
Returns the non-directed graph formed through the sequence
}
\description{
A function that from a biological sequence generates a graph not addressed having as words vertices, this being able to have its size parameter set by the' word 'parameter. The connections between words depend of the' step 'parameter that indicates the next connection to be formed
}
\author{
Eric Augusto Ito
}
