% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures}
\alias{measures}
\title{Abstracting Characteristics on Network Structure}
\usage{
measures(graph)
}
\arguments{
\item{graph}{The graph that will be measured}
}
\value{
Vector with the results of the measurements in order:
Average path length, cluster coefficient, degree, assortativity,
betweenness, standard deviation, maximum, minimum, number of motifs
size 3 and number of motifs of size 4
}
\description{
Given a graph, it is made up of several
features on the graph structure and returns a vector
with the data obtained
}
\details{

}
\examples{

}
\seealso{

}
\author{
Eric Augusto Ito
}
