% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minMax.R
\name{minMax}
\alias{minMax}
\title{Minimum and maximum}
\usage{
minMax(matrix, mRNA, lncRNA, sncRNA, rangeMinMax)
}
\arguments{
\item{matrix}{Array with results numerics}

\item{mRNA}{Integer number of mRNA sequences}

\item{lncRNA}{Integer number of lncRNA sequences}

\item{sncRNA}{Integer number of sncRNA sequences}

\item{rangeMinMax}{Vector that will be returned with the minimum and maximum values}
}
\value{
Returns the vector with the minimum and maximum values for the scale
}
\description{
Verifies the minimum and maximum values of the results.
}
\author{
Eric Augusto Ito
}
