% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeq.R
\name{readSeq}
\alias{readSeq}
\alias{lerSeq}
\title{Returns a sequence of a data set of type FASTA}
\usage{
lerSeq(arqFasta, x)
}
\arguments{
\item{arqFasta}{FASTA type file containing the sequences}

\item{x}{Integer that indicates which sequence will be returned}
}
\value{
Returns the specified string
}
\description{
Given a set of strings in a FASTA file and
given an integer X, the sequence X of the set will be returned
of data
}
\details{

}
\examples{


}
\seealso{

}
\author{
Eric Augusto Ito
}
