% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Applies threshold on the network from a value}
\usage{
threshold(x, net)
}
\arguments{
\item{x}{Value that would limit the edges}

\item{net}{Network where the edges will be cut}
}
\value{
Returns the complex network with the cuts already made
}
\description{
Given an integer value X, a cut, that is, edges that are cut will be assigned zero. This cut will be done in the network where the edges have a weight less than the value of X.
}
\author{
Eric Augusto Ito
}
