% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraph2D.R
\name{createGraph2D}
\alias{createGraph2D}
\title{Creates a two-dimensional graph between a measure and the threshold}
\usage{
createGraph2D(matrix, numSeqMRNA, numSeqLNCRNA, nameMeasure)
}
\arguments{
\item{matrix}{matrix of the measure for the creation of two-dimensional graph}

\item{numSeqMRNA}{Integer number of mRNA sequences}

\item{numSeqLNCRNA}{Integer number of lncRNA sequences}

\item{nameMeasure}{Character Parameter that defines the name of the measure to put in the title of the graph}
}
\description{
For an analysis of each measure, the createGraph2D () function was created in order to visualize the behavior of each measurement in relation to the threshold. This function creates a graph (Measure x Threshold) from an array, mRNA sequences are given the blue color, the lncRNA sequences are given a red color. In cases where there is a third class this will be given the green color
}
\author{
Eric Augusto Ito
}
