% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraph3D.R
\name{createGraph3D}
\alias{createGraph3D}
\title{Creates an untargeted graph from a biological sequence}
\usage{
createGraph3D(matrix1, matrix2, numSeqMRNA, numSeqLNCRNA, nameMeasure1,
  nameMeasure2)
}
\arguments{
\item{matrix1}{matrix of the first measure for the creation of the three-dimensional chart}

\item{matrix2}{matrix of the second measure for the creation of the three-dimensional chart}

\item{numSeqMRNA}{number of mRNA sequences}

\item{numSeqLNCRNA}{number of lncRNA sequences}

\item{nameMeasure1}{name of the first measure to put in the title of the graph}

\item{nameMeasure2}{name of the second measure to put in the title of the graph}
}
\value{
Returns the non-directed graph formed through the sequence
}
\description{
A function that from a biological sequence generates a graph
not addressed having as words vertices, this being able to have its size
parameter set by the' word 'parameter. The connections between words depend
of the' step 'parameter that indicates the next connection to be formed
}
\details{

}
\examples{

}
\seealso{

}
\author{
Eric Augusto Ito
}
