% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures}
\alias{measures}
\title{Abstracting Characteristics on Network Structure}
\usage{
measures(graph)
}
\arguments{
\item{graph}{The complex network that will be measured}
}
\value{
Return a vector with the results of the measurements in order:
Average shortest path length, clustering Coefficient, degree, assortativity,
betweenness, standard deviation, maximum, minimum, number of motifs
size 3 and number of motifs of size 4
}
\description{
Given a graph, it is made up of several features on the graph structure and returns a vector with the data obtained
}
\author{
Eric Augusto Ito
}
