% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reschedule.R
\name{reschedule}
\alias{reschedule}
\alias{reescalar}
\title{rescales the results between values ​​from 0 to 1}
\usage{
reescalar(matrix, mRNA, lncRNA, sncRNA)
}
\arguments{
\item{matrix}{array with results}

\item{mRNA}{Number of mRNA sequences}

\item{lncRNA}{Number of lncRNA sequences}

\item{sncRNA}{Number of sncRNA sequences}
}
\value{
Returns the array with the rescaled values
}
\description{
Given the results the data is rescaled
for values ​​between 0 and 1, so that the length of the sequences does not
influence the results. The rescaling of the mRNA and lncRNA
are made separately
}
\details{

}
\examples{

}
\seealso{

}
\author{
Eric Augusto Ito
}
